+++ lib/smart_core/ext.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# @api private
+# @since 0.9.0
+module SmartCore::Ext
+  require_relative 'ext/basic_object_as_object'
+end